engine/resolution_engine/io.py
"""
Input/Output Module

# config, logging, audit ledger
"""

import json
import yaml
import logging
from typing import Dict, Any, List, Optional
from pathlib import Path
from datetime import datetime


class ConfigManager:
    """
    Configuration management
    Coming soon - complete config implementation
    """
    
    def __init__(self, config_path: Optional[str] = None):
        """Initialize configuration manager"""
        self.config_path = config_path
        self.config = {}
        # TODO: Load configuration
        pass
    
    def load_config(self, path: str) -> Dict[str, Any]:
        """
        Load configuration from YAML file
        Coming soon - config loading implementation
        """
        # TODO: Implement YAML config loading
        pass
    
    def save_config(self, config: Dict[str, Any], path: str):
        """
        Save configuration to YAML file
        Coming soon - config saving implementation
        """
        # TODO: Implement config saving
        pass


class Logger:
    """
    System logging functionality
    Coming soon - complete logging implementation
    """
    
    def __init__(self, log_level: str = 'INFO', log_file: Optional[str] = None):
        """Initialize logger"""
        self.logger = logging.getLogger('resolution_engine')
        # TODO: Configure logging
        pass
    
    def log_state_transition(self, from_state: Any, to_state: Any, action: Any):
        """
        Log state transitions
        Coming soon - state transition logging
        """
        # TODO: Implement state transition logging
        pass
    
    def log_action(self, action: Any, context: Dict[str, Any]):
        """
        Log actions taken
        Coming soon - action logging
        """
        # TODO: Implement action logging
        pass


class AuditLedger:
    """
    Audit trail and compliance logging
    Coming soon - complete audit implementation
    """
    
    def __init__(self, ledger_path: str):
        """Initialize audit ledger"""
        self.ledger_path = Path(ledger_path)
        # TODO: Initialize audit ledger
        pass
    
    def record_decision(self, decision: Dict[str, Any], context: Dict[str, Any]):
        """
        Record decision in audit ledger
        Coming soon - decision recording implementation
        """
        # TODO: Implement decision recording
        pass
    
    def record_outcome(self, outcome: Dict[str, Any], decision_id: str):
        """
        Record outcome in audit ledger
        Coming soon - outcome recording implementation
        """
        # TODO: Implement outcome recording
        pass
