engine/requirements.txt
# Resolution Engine Requirements

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine learning
scikit-learn>=1.0.0

# Configuration and CLI
pyyaml>=6.0
click>=8.0.0

# Logging and monitoring
structlog>=21.0.0

# Development dependencies (optional)
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0
